making improvements
changed line 386 in dense.py;
moved some helper functions to test_dense_customized;
removed unused helper functions;
